requirements.txt
pandas==2.2.1 # Automatically Installs NumPy
boto3==1.34.82
requests==2.31.0
python-dotenv==1.0.1
pyyaml==6.0.1